Remove DTO-dependency in SPKTjenestepensjonServicePre2025

+++ src/main/kotlin/no/nav/pensjon/simulator/tjenestepensjon/pre2025/api/acl/v1/SimulerOffentligTjenestepensjonMapperV1.kt
@@ -79,7 +79,7 @@ object SimulerOffentligTjenestepensjonMapperV1 {
         Opptjeningsperiode(
             fom = dto.datoFom,
             tom = dto.datoTom,
-            stillingsprosent = dto.stillingsprosent,
+            stillingsprosent = dto.stillingsprosent.toDouble(),
             aldersgrense = dto.aldersgrense,
             faktiskHovedloenn = dto.faktiskHovedlonn,
             stillingsuavhengigTilleggsloenn = dto.stillingsuavhengigTilleggslonn
